+++ docs/recipes/automation/tp-setup/bootstrap/page_object/po_user_management.py
@@ -1,11 +1,13 @@
 #  Copyright (c) 2025. Cloud Software Group, Inc. All Rights Reserved. Confidential & Proprietary
-
+from page_object.po_global import PageObjectGlobal
 from utils.color_logger import ColorLogger
 from utils.env import ENV
 from utils.report import ReportYaml
+from utils.util import Util
 
-class PageObjectUserManagement:
+class PageObjectUserManagement(PageObjectGlobal):
     def __init__(self, page):
+        super().__init__(page)
         self.page = page
         self.env = ENV
 
@@ -22,9 +24,8 @@ class PageObjectUserManagement:
 
                 print("Grant permission for " + permission)
 
-    # TODO: remove this function after bug: PCP-11136 is fixed
-    def set_user_permission_for_bug(self):
-        print("For bug: PCP-11136, Start set user permission...")
+    def goto_assign_permissions(self):
+        print("Start set user permission...")
         self.page.click("#nav-bar-menu-item-usrMgmt")
         self.page.click("#users-menu-item")
         self.page.locator(f'.user-name-text[id="go-to-user-details-{ENV.DP_USER_EMAIL}"]').wait_for(state="visible")
@@ -36,29 +37,15 @@ class PageObjectUserManagement:
         self.page.locator(".pl-dropdown-menu__action", has_text="Update permissions").click()
         print(f"Clicked 'Update permissions' from dropdown list")
 
-        self.page.wait_for_timeout(1000)
-        print(f"Assign Permissions for {ENV.DP_USER_EMAIL}")
-
-        self.grant_permission("Data plane Manager")
-        self.grant_permission("Capability Manager")
-        self.grant_permission("Application Manager")
-
-        if not self.page.locator("#next-assign-permissions").is_disabled():
-            self.page.click("#next-assign-permissions")
-            self.page.click("#assign-permissions-update")
-            ColorLogger.success(f"Grant permission to {ENV.DP_USER_EMAIL}")
-        else:
-            ColorLogger.warning(f"Nothing changed for user permission.")
-
-        print("For bug: PCP-11136, End set user permission...")
+        self.page.locator(".policy-selector-container").wait_for(state="visible")
+        print("Assign permissions page is loaded.")
 
     def set_user_permission(self):
         if ReportYaml.get(".ENV.REPORT_USER_PERMISSION") == "true":
             ColorLogger.success(f"In {ENV.TP_AUTO_REPORT_YAML_FILE} file, user permission is already set.")
             return
         ColorLogger.info("Setting user permission...")
-        self.page.locator("#nav-bar-menu-list-dataPlanes").wait_for(state="visible")
-        self.page.click("#nav-bar-menu-list-dataPlanes")
+        self.goto_left_navbar("Data Planes")
         self.page.locator("#register-dp-button").wait_for(state="visible")
         print("Checking if user has permission...")
         if not self.page.locator("#register-dp-button").is_disabled():
@@ -66,33 +53,7 @@ class PageObjectUserManagement:
             ReportYaml.set(".ENV.REPORT_USER_PERMISSION", True)
             return
 
-        # self.set_user_permission_for_bug()
-
-        print("Start set user permission...")
-        self.page.click("#nav-bar-menu-item-usrMgmt")
-        self.page.click("#users-menu-item")
-        self.page.locator(f'.user-name-text[id="go-to-user-details-{ENV.DP_USER_EMAIL}"]').wait_for(state="visible")
-        print(f"{ENV.DP_USER_EMAIL} is found.")
-        # check if user has all permissions
-        # self.page.locator(f'.user-name-text[id="go-to-user-details-{ENV.DP_USER_EMAIL}"]').click()
-
-        # self.page.wait_for_timeout(1000)
-        # print(f"Assign Permissions for {ENV.DP_USER_EMAIL}")
-        # # if table.permissions has 8 rows, then exit this function
-        # if self.page.locator("table.permissions .pl-table__row").count() == 8:
-        #     ColorLogger.success(f"User {ENV.DP_USER_EMAIL} already has all permissions.")
-        #     ReportYaml.set(".ENV.REPORT_USER_PERMISSION", True)
-        #     return
-
-        # self.page.click("#assign-permissions-btn")
-
-        self.page.locator("team-members tr", has=self.page.locator(f'.user-name-text[id="go-to-user-details-{ENV.DP_USER_EMAIL}"]')).locator("dropdown-button button#changeME-dropdown-label").click()
-        print(f"Click on dropdown button for user {ENV.DP_USER_EMAIL}")
-        self.page.locator(".pl-dropdown-menu__action", has_text="Update permissions").wait_for(state="visible")
-        self.page.locator(".pl-dropdown-menu__action", has_text="Update permissions").click()
-        print(f"Clicked 'Update permissions' from dropdown list")
-
-        self.page.wait_for_timeout(1000)
+        self.goto_assign_permissions()
         print(f"Assign Permissions for {ENV.DP_USER_EMAIL}")
         # if it has 8 green icons, then exit this function
         if self.page.locator(".policy-selector-container .green-check-icon").count() >= 8:
@@ -116,3 +77,54 @@ class PageObjectUserManagement:
         else:
             ColorLogger.success(f"User {ENV.DP_USER_EMAIL} already has all permissions.")
             ReportYaml.set(".ENV.REPORT_USER_PERMISSION", True)
+
+    def grant_product_permission(self, dp_name, app_name):
+        permission = "Product Permission"
+        ColorLogger.info(f"Granting  {permission} for "+ dp_name + " => " + app_name)
+        self.goto_assign_permissions()
+        print("Checking if product permission is visible...")
+        if Util.check_dom_visibility(self.page, self.page.locator(".policy-description", has_text=permission), 2, 4):
+            print("Product permission is visible.")
+            self.page.locator(".policy-description", has_text=permission).click()
+            print("Checking product permission for " + dp_name + " => " + app_name)
+
+            dataplane_selector = self.page.locator('.dataplane-selector', has=self.page.locator('.dp-name-text', has_text=dp_name))
+            if Util.check_dom_visibility(self.page, dataplane_selector, 2, 4):
+                print("Dataplane " + dp_name + " is found in product permission list.")
+                dataplane_selector.click()
+                print("Clicked on dataplane " + dp_name)
+
+                app_selector = dataplane_selector.locator(".products-container .product-item", has_text=app_name)
+                if Util.check_dom_visibility(self.page, app_selector, 2, 4):
+                    print("Product " + dp_name + " => " + app_name + " is found in product permission list.")
+                    app_selector.click()
+                    print("Clicked on product " + dp_name + " => " + app_name)
+
+                    is_all_domain_checked = self.page.locator(".domain-selector-container .wildcard-domains-checkbox input").get_attribute("aria-checked")
+                    if is_all_domain_checked != "true":
+                        self.page.locator(".domain-selector-container .wildcard-domains-checkbox label", has_text="All current and future domains").click()
+                        is_all_domain_checked = "true"
+                        print("Clicked on 'All current and future domains' for " + dp_name + " => " + app_name)
+
+                    is_write_selected = self.page.locator(".domains-table-container .pl-table__header input[name='allDomainsWriteSelected']").get_attribute("aria-checked")
+                    if is_write_selected != "true":
+                        self.page.locator(".domains-table-container .pl-table__header label", has_text="Write").click()
+                        is_write_selected = "true"
+                        print("Clicked on 'Write' permission for " + dp_name + " => " + app_name)
+
+                    print("All current and future domains: " + is_all_domain_checked)
+                    print("Write permission: " + is_write_selected)
+
+                    if not self.page.locator("#next-assign-permissions").is_disabled():
+                        self.page.click("#next-assign-permissions")
+                        self.page.click("#assign-permissions-update")
+                        ColorLogger.success(f"Grant All permission to {ENV.DP_USER_EMAIL}")
+                        ReportYaml.set(".ENV.REPORT_USER_PERMISSION", True)
+                    else:
+                        ColorLogger.warning("Next button is disabled, nothing to update.")
+                else:
+                    ColorLogger.warning(f"Product {dp_name} => {app_name} is not found in product permission list.")
+            else:
+                ColorLogger.warning(f"Dataplane {dp_name} is not found in product permission list.")
+        else:
+            ColorLogger.info("Product permission is not visible, skipping...")